Replace ExposureNotificatonEventSubject to Prism mechanism.
