+++ Covid19Radar/Covid19Radar/IExposureNotificationEventCallback.cs
@@ -1,0 +1,10 @@
+﻿namespace Covid19Radar
+{
+    public interface IExposureNotificationEventCallback
+    {
+        public void OnEnabled() { }
+        public void OnDeclined() { }
+        public void OnGetTekHistoryAllowed() { }
+        public void OnPreauthorizeAllowed() { }
+    }
+}
